Move Code to gladius Core
