+++ gladius_core/Cargo.toml
@@ -1,25 +1,9 @@
 [package]
-name = "gladius_slicer"
-version = "0.4.1"
-authors = ["mrhatman <lucasince@gmail.com>"]
+name = "gladius_core"
+version = "0.1.0"
 edition = "2021"
 
-[profile.release]
-lto = true
-
-[workspace]
-members = [ "gladius_core", 
-    "gladius_shared"
-]
-
-[features]
-json_schema_gen = ["gladius_shared/json_schema_gen"]
-[lints.clippy]
-allow-unwrap-in-tests = "allow"
-
 [dependencies]
-clap = { version = "4", features = ["derive"]}
-simple_logger = "5.0.0"
 log = "0.4.22"
 
 nalgebra = "0.33"
@@ -40,7 +24,6 @@ rand = "0.8.5"
 
 evalexpr = "12.0.1"
 
-gladius_shared = { path = "gladius_shared", version = "*" }
-gladius_core = { path = "gladius_core", version = "*" }
-
 binary-heap-plus = "0.5.0"
+
+gladius_shared = { path = "../gladius_shared", version = "*" }
