+++ gladius_core/src/optimizer.rs
