+++ gladius_core/src/plotter/mod.rs
@@ -8,13 +8,12 @@ pub(crate) mod support;
 pub use crate::plotter::infill::*;
 use crate::plotter::perimeter::inset_polygon_recursive;
 use crate::plotter::polygon_operations::PolygonOperations;
-use crate::utils::point_lerp;
-use crate::{Object, Settings, StateChange};
 use coordinate_position::CoordPos;
 use geo::prelude::*;
 use geo::{Coord, MultiPolygon, Polygon};
-use gladius_shared::settings::SkirtSettings;
-use gladius_shared::types::{Command, Move, MoveChain, MoveType, RetractionType, Slice};
+use gladius_shared::settings::{Settings, SkirtSettings};
+use gladius_shared::types::{Command, Move, MoveChain, MoveType, Object, RetractionType, Slice, StateChange};
+use gladius_shared::utils::point_lerp;
 use itertools::Itertools;
 use log::info;
 use ordered_float::OrderedFloat;
