+++ gladius_core/src/plotter/monotone.rs
@@ -1,4 +1,4 @@
-use crate::utils::{orientation, Orientation};
+use gladius_shared::utils::{orientation, Orientation};
 use geo::{Coord, Polygon, SimplifyVwPreserve};
 use geo_svg::{Color, ToSvg};
 use itertools::Itertools;
