+++ gladius_core/src/plotter/polygon_operations.rs
